:recycle: Refactored file structure and outsource premium and ios validations on dashboard view into seperate functions
