+++ music-app/src/views/Dashboard.js
@@ -8,15 +8,9 @@ import List from '@material-ui/core/List';
 import { makeStyles } from '@material-ui/core/styles';
 
 import { getlikedSongs, getUsers, getSpotifyAccountDetails } from '../actions';
-import LikedSongs from './LikedSongs.js';
-import MusicPlayer from './MusicPlayer.js';
+import LikedSongs from '../components/LikedSongs.js';
+import MusicPlayer from '../components/MusicPlayer.js';
 import '../App.css';
-
-// import Fab from '@material-ui/core/Fab';
-// import MenuIcon from '@material-ui/icons/Menu';
-// import AddIcon from '@material-ui/icons/Add';
-// import SearchIcon from '@material-ui/icons/Search';
-// import MoreIcon from '@material-ui/icons/MoreVert';
 
 const useStyles = makeStyles(theme => ({
   button: {
@@ -31,15 +25,24 @@ class Dashboard extends React.Component {
     localStorage.removeItem('token');
     this.props.history.push('/logout');
   };
-  render() {
-    if (
-      (this.props.spotifyUser.product &&
-        this.props.fetchingSpotifyUser === false &&
-        this.props.spotifyUser.product !== 'premium') ||
-      window.navigator.platform === 'iPhone' ||
+
+  checkPremiumUser = () => {
+    return this.props.spotifyUser.product &&
+      this.props.fetchingSpotifyUser === false &&
+      this.props.spotifyUser.product !== 'premium'
+      ? true
+      : false;
+  };
+
+  checkNoIOS = () => {
+    return window.navigator.platform === 'iPhone' ||
       window.navigator.platform === 'iPad' ||
       window.navigator.platform === 'iPod'
-    ) {
+      ? true
+      : false;
+  };
+  render() {
+    if (this.checkPremiumUser() || this.checkNoIOS()) {
       this.props.history.push('/info');
     }
     return (
