+++ music-app/src/views/Info.js
